refactor: 알림 관련 컴포넌트 분리
