+++ features/notifications/components/NotificationPermissionModal/IOSSafariModal.tsx
@@ -1,0 +1,77 @@
+'use client';
+
+import {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+  DialogDescription,
+} from '@/shared/components/ui/dialog';
+import { Button } from '@/shared/components/ui/button';
+import { Bell, Share, PlusSquare } from 'lucide-react';
+
+interface IOSSafariModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+}
+
+export default function IOSSafariModal({ isOpen, onClose }: IOSSafariModalProps) {
+  return (
+    <Dialog open={isOpen} onOpenChange={(open) => !open && onClose()}>
+      <DialogContent className="sm:max-w-md">
+        <DialogHeader className="text-center">
+          <div className="mx-auto w-16 h-16 bg-brand-100 dark:bg-brand-900/30 rounded-full flex items-center justify-center mb-4">
+            <Bell className="w-8 h-8 text-brand-500" />
+          </div>
+          <DialogTitle className="text-xl">푸시 알림 받기</DialogTitle>
+          <DialogDescription className="text-slate-500 dark:text-slate-400">
+            iOS에서 푸시 알림을 받으려면 앱을 홈 화면에 추가해주세요
+          </DialogDescription>
+        </DialogHeader>
+
+        <div className="space-y-4 my-4">
+          <div className="p-4 bg-slate-50 dark:bg-slate-800/50 rounded-xl space-y-3">
+            <p className="text-sm font-medium text-slate-700 dark:text-slate-200">
+              홈 화면에 추가하는 방법
+            </p>
+            <div className="flex items-center gap-3">
+              <div className="w-8 h-8 bg-brand-100 dark:bg-brand-900/30 rounded-lg flex items-center justify-center shrink-0">
+                <Share className="w-4 h-4 text-brand-500" />
+              </div>
+              <span className="text-sm text-slate-600 dark:text-slate-300">
+                1. 하단의 <strong>공유</strong> 버튼을 탭하세요
+              </span>
+            </div>
+            <div className="flex items-center gap-3">
+              <div className="w-8 h-8 bg-brand-100 dark:bg-brand-900/30 rounded-lg flex items-center justify-center shrink-0">
+                <PlusSquare className="w-4 h-4 text-brand-500" />
+              </div>
+              <span className="text-sm text-slate-600 dark:text-slate-300">
+                2. <strong>홈 화면에 추가</strong>를 선택하세요
+              </span>
+            </div>
+          </div>
+
+          <div className="space-y-2">
+            <p className="text-xs text-slate-500 dark:text-slate-400 text-center">
+              홈 화면에 추가하면 이런 알림을 받을 수 있어요
+            </p>
+            <div className="flex gap-2 justify-center text-xs text-slate-500">
+              <span className="px-2 py-1 bg-slate-100 dark:bg-slate-800 rounded">면접 리마인더</span>
+              <span className="px-2 py-1 bg-slate-100 dark:bg-slate-800 rounded">서류 마감</span>
+              <span className="px-2 py-1 bg-slate-100 dark:bg-slate-800 rounded">현황 업데이트</span>
+            </div>
+          </div>
+        </div>
+
+        <Button
+          variant="default"
+          className="w-full font-bold"
+          onClick={onClose}
+        >
+          확인
+        </Button>
+      </DialogContent>
+    </Dialog>
+  );
+}
